Add explicit_math option to posts.
By default, I want $ $ to render math; I decided that `$ $` was too
verbose. But I have some posts that that will interfere with, so I need
an option to turn it off.

IIRC, queueing 'Typeset' on an ignored element doesn't do anything, so I
need to make code tags not-ignored when I'm typesetting them, regardless
of explicit_math. But with explicit_math off, I need to make sure
they're ignored when I'm not typesetting them.

Will probably need to play around with this to make it work...