Remove downscaling of math.
I dunno why it's a good size now by default, but it does seem to be.
Maybe still a little big, there's a thing somewhere setting it to 122%
which would mean scaling to 82% might be better? But I'm not sure it
looks better than the current size when I try it; and there's value in
simplicity.